third_party/blink/web_tests/fast/forms/required-attribute-002-expected.txt

There are three form control elements below, all optional. They're set as required via required DOM attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Before set:
PASS requiredFor("input") is false
PASS requiredFor("textarea") is false
PASS requiredFor("select") is false


After set:
PASS requiredFor("input") is true
PASS requiredFor("textarea") is true
PASS requiredFor("select") is true


PASS successfullyParsed is true

TEST COMPLETE

  
X
